requirements.txt
# Core dependencies
asyncio
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
alembic==1.13.1

# Telegram
aiogram==3.4.1
telethon==1.34.0

# OpenAI
openai==1.6.1
tiktoken==0.5.2

# Data processing
pydantic==2.5.3
pydantic-settings==2.1.0
python-dateutil==2.8.2

# Async support
aiofiles==23.2.1
aiohttp

# Logging
loguru==0.7.2

# Environment
python-dotenv==1.0.0

# Security
cryptography

# Utilities
jinja2==3.1.2
socks
rich==13.7.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.2
black==23.12.1
isort==5.13.2